gamma mapping extended to 16bit - fixes #425 and #424 (only if opencv is available)

+++ ImageLounge/src/DkMath.cpp
@@ -251,7 +251,7 @@ double DkRotatingRect::getAngle() const {
 
 float DkRotatingRect::getAngleDeg() const {
 
-	double sAngle = (getAngle())*DK_RAD2DEG;
+	float sAngle = (float)(getAngle()*DK_RAD2DEG);
 
 	while (sAngle > 90)
 		sAngle -= 180;
